exceptions

+++ src/Actions/ResetPassword.php
@@ -5,6 +5,7 @@ namespace Lab19\Cart\Actions;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
+use Lab19\Cart\Exceptions\GernzyException;
 use Lab19\Cart\Models\PasswordResets;
 use Lab19\Cart\Models\User;
 
@@ -13,14 +14,24 @@ class ResetPassword
     public static function handle($args): User
     {
         // Get the records from the cart_password_resets table and compute the difference in time to see if token has expired
-        $resetRecord =  PasswordResets::where('email', $args['email'])->firstOrFail();
+        $resetRecord =  PasswordResets::where('email', $args['email'])->first();
+        if ($resetRecord === null) {
+            throw new GernzyException(
+                'The provided email does not exist.',
+                'Please resubmit a password reset request.'
+            );
+        }
+
         $createdAtTime = Carbon::parse($resetRecord->created_at);
         $timeDiff = $createdAtTime->diffInHours(Carbon::now());
 
         // Check if created less than 24 hours ago
         if ($timeDiff > 24) {
             $resetRecord->delete();
-            return false;
+            throw new GernzyException(
+                'Token expired',
+                'The token has expired, please resubmit a password reset request.'
+            );
         }
 
         // Check the record and compare the token from the args to the one in the table and return email, then delete the record
